Move test to avoid extra dependency
